+++ src/endpoints/authentication/cert_auth.cpp
@@ -5,6 +5,7 @@
 
 #include "ccf/ds/locking.h"
 #include "ccf/ds/x509_time_fmt.h"
+#include "ccf/node_context.h"
 #include "ccf/rpc_context.h"
 #include "ccf/service/tables/members.h"
 #include "ccf/service/tables/nodes.h"
@@ -236,6 +237,33 @@ namespace ccf
 
     auto identity = std::make_unique<AnyCertAuthnIdentity>();
     identity->cert = caller_cert;
+    return identity;
+  }
+
+  SelfCertAuthnPolicy::SelfCertAuthnPolicy(
+    const AbstractNodeContext& node_context_) :
+    node_context(node_context_)
+  {}
+
+  std::unique_ptr<AuthnIdentity> SelfCertAuthnPolicy::authenticate(
+    ccf::kv::ReadOnlyTx& tx,
+    const std::shared_ptr<ccf::RpcContext>& ctx,
+    std::string& error_reason)
+  {
+    auto identity = AnyCertAuthnPolicy::authenticate(tx, ctx, error_reason);
+    if (identity == nullptr)
+    {
+      return nullptr;
+    }
+
+    const auto caller_node_id =
+      compute_node_id_from_cert_der(ctx->get_session_context()->caller_cert);
+    if (caller_node_id != node_context.get_node_id())
+    {
+      error_reason = "Only the node itself can call this endpoint.";
+      return nullptr;
+    }
+
     return identity;
   }
 }